chore: rename pop-up
