+++ packages/desktop/lib/auxiliary/popup/enums/popup-id.enum.ts
@@ -17,6 +17,7 @@ export enum PopupId {
     Diagnostics = 'diagnostics',
     EnableLedgerBlindSigning = 'enableLedgerBlindSigning',
     ErrorLog = 'errorLog',
+    EvmTransactionFromDapp = 'EvmTransactionFromDapp',
     FaucetRequest = 'faucetRequest',
     ImportErc20Token = 'importErc20Token',
     LegalUpdate = 'legalUpdate',
@@ -32,7 +33,6 @@ export enum PopupId {
     RemoveProposal = 'removeProposal',
     Revote = 'revote',
     SendFlow = 'sendFlow',
-    SignEvmTransaction = 'signEvmTransaction',
     SignMessage = 'signMessage',
     StopVoting = 'stopVoting',
     SyncAccounts = 'syncAccounts',
