+++ packages/shared/src/lib/auxiliary/wallet-connect/handlers/eth_transaction.handler.ts
@@ -45,7 +45,7 @@ export async function handleEthTransaction(
     }
 
     openPopup({
-        id: PopupId.SignEvmTransaction,
+        id: PopupId.EvmTransactionFromDapp,
         props: {
             account,
             chain,
